Move repositiory-related postfactorymenu from meeting to repository module
