+++ opengever/meeting/menu.py
@@ -1,7 +1,6 @@
 from five import grok
 from opengever.base.menu import FilteredPostFactoryMenu
 from opengever.meeting.committee import ICommittee
-from opengever.repository.interfaces import IRepositoryFolder
 from zope.interface import Interface
 
 
@@ -18,12 +17,3 @@ class CommitteePostFactoryMenu(FilteredPostFactoryMenu):
             return True
 
         return False
-
-
-class RepositoryFolderPostFactoryMenu(FilteredPostFactoryMenu):
-    grok.adapts(IRepositoryFolder, Interface)
-
-    def is_filtered(self, factory):
-        factory_id = factory.get('id')
-        if factory_id == u'opengever.meeting.meetingdossier':
-            return True
